Refactor: move auth files into separate folder
